@@ -32,7 +32,7 @@ class GroupsController < ApplicationController
   # POST /groups
   # POST /groups.json
   def create
-    @group = Group.new(params[:group])
+    @group = Group.new(create_params)
     @group[:owner_id] = current_user.id
     afterSave(@group) if @group.save
     respond_with(@group)
@@ -42,7 +42,7 @@ class GroupsController < ApplicationController
   # PUT /groups/1.json
   def update
     @group = current_user.own_groups.find(params[:id])
-    flash[:notice] = 'Group was successfully updated.' if @group.update_attributes(params[:group])
+    flash[:notice] = 'Group was successfully updated.' if @group.update_attributes(update_params)
     respond_with(@group)
   end
 
@@ -50,6 +50,16 @@ class GroupsController < ApplicationController
   def afterSave(group)
     flash[:notice] = 'Group was successfully created.' 
     GroupTopic.create(topic_id:Topic.getFirstTopic.id, group_id:group.id)
+  end
+  
+  private
+  def create_params
+    params.require(:group).permit(:name, :description, :visible, :photo)
+  end
+  
+  private
+  def update_params
+    params.require(:group).permit(:name, :description, :visible, :photo)
   end
 
 end
