@@ -1,5 +1,7 @@
 class Users::RegistrationsController < Devise::RegistrationsController
- 
+  before_filter :configure_permitted_parameters
+
+
   def new
     super
   end
@@ -15,4 +17,16 @@ class Users::RegistrationsController < Devise::RegistrationsController
     hash[:uid] = User.create_unique_string
     super
   end
+  
+  protected
+  # my custom fields are :name, :heard_how
+  def configure_permitted_parameters
+    devise_parameter_sanitizer.for(:sign_up) do |u|
+      u.permit(:username, :email, :locale_id, :password, :password_confirmation)
+    end
+    devise_parameter_sanitizer.for(:account_update) do |u|
+      u.permit(:username, :email, :locale_id, :password, :password_confirmation, :current_password, :photo)
+    end
+  end
+  
 end
